+++ ballcat-starters/ballcat-spring-boot-starter-redis/src/main/java/com/hccake/ballcat/common/redis/config/RedisAutoConfiguration.java
@@ -7,7 +7,7 @@ import com.hccake.ballcat.common.redis.serialize.CacheSerializer;
 import com.hccake.ballcat.common.redis.serialize.JacksonSerializer;
 import com.hccake.ballcat.common.redis.serialize.PrefixJdkRedisSerializer;
 import com.hccake.ballcat.common.redis.serialize.PrefixStringRedisSerializer;
-import com.hccake.ballcat.common.redis.RedisHelpers;
+import com.hccake.ballcat.common.redis.RedisHelper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
@@ -99,10 +99,10 @@ public class RedisAutoConfiguration {
 	}
 
 	@Bean(name = "com.hccake.ballcat.common.redis.RedisHelpers")
-	@ConditionalOnMissingBean(RedisHelpers.class)
-	public RedisHelpers redis(StringRedisTemplate template) {
-		RedisHelpers.setTemplate(template);
-		return new RedisHelpers();
+	@ConditionalOnMissingBean(RedisHelper.class)
+	public RedisHelper redis(StringRedisTemplate template) {
+		RedisHelper.setTemplate(template);
+		return new RedisHelper();
 	}
 
 }